WIP test_schemas fixing"
