+++ examples/app.py
@@ -75,6 +75,8 @@ from invenio_pidrelations.contrib.versioning import PIDNodeVersioning, \
 from invenio_pidrelations.indexers import index_relations
 from invenio_pidrelations.models import PIDRelation
 from invenio_pidrelations.utils import resolve_relation_type_config
+from test_helpers import PIDRelationsMixin
+
 
 # Create Flask application
 app = Flask(__name__, template_folder='.')
@@ -100,7 +102,7 @@ record_resolver = Resolver(
 )
 
 
-class SimpleRecordSchema(Schema):
+class SimpleRecordSchema(Schema, PIDRelationsMixin):
     """Tiny schema for our simple record."""
 
     recid = fields.Str()
